--- examples/configs/homography_estimation.yaml
@@ -1,33 +1,0 @@
-seed: 0
-use_gpu: True
-
-imgH: 160
-imgW: 200
-#imgH: 20
-#imgW: 25
-photo_aug: True
-
-train_chkpt: None
-eval_chkpt: None
-# train_chkpt: "/home/joe/projects/mpSLAM/theseus/examples/outputs/chkpts/epoch_001_step_1800.pt"
-# eval_chkpt: "/home/joe/projects/mpSLAM/theseus/examples/outputs/chkpts/epoch_001_step_1800.pt"
-
-inner_optim:
-  max_iters: 25
-  step_size: 0.1
-  backward_mode: "implicit"
-  reg_w: 1e-2
-
-outer_optim:
-  lr: 1e-5
-  num_epochs: 10
-
-#train_batch_size: 5
-#test_max_iters: 50
-#test_batch_size: 5
-train_batch_size: 1
-test_max_iters: 50
-test_batch_size: 1
-
-save_gifs: True
-save_model: True
